+++ packages/toolpad-app/src/runtime/toolpadComponents/template.ts
@@ -1,7 +1,7 @@
 import { ArgTypeDefinitions, PropValueType } from '@mui/toolpad-core';
 
-import * as appDom from '../appDom';
-import { PageViewState } from '../types';
+import * as appDom from '../../appDom';
+import { PageViewState } from '../../types';
 
 export const isTemplateDescendant = (
   dom: appDom.AppDom,
