+++ packages/toolpad-app/src/toolpad/AppEditor/PageEditor/NodeAttributeEditor.tsx
@@ -16,7 +16,7 @@ import BindableEditor from './BindableEditor';
 import { usePageEditorState } from './PageEditorProvider';
 import { getDefaultControl } from '../../propertyControls';
 import MarkdownTooltip from '../../../components/MarkdownTooltip';
-import { isTemplateDescendant } from '../../../toolpadComponents/template';
+import { isTemplateDescendant } from '../../../runtime/toolpadComponents/template';
 import { NON_BINDABLE_CONTROL_TYPES } from '../../../runtime/constants';
 
 export interface NodeAttributeEditorProps<P extends object, K extends keyof P = keyof P> {
