+++ packages/toolpad-app/src/types.ts
@@ -182,3 +182,7 @@ export interface RuntimeState {
   // We start out with just the rendertree. The ultimate goal will be to move things out of this tree
   dom: appDom.RenderTree;
 }
+
+export interface AppCanvasState extends RuntimeState {
+  savedNodes: NodeHashes;
+}
